infrastructure/gcp/functions/trigger_price_history_backup/requirements.txt
functions-framework==3.*
google-cloud-firestore
google-auth
